+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/TreeDiff.hs
@@ -14,7 +14,7 @@ module Test.Cardano.Ledger.Conway.TreeDiff (
 
 import Cardano.Ledger.Alonzo.Plutus.Context (ContextError)
 import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.Conway (ConwayEra)
+import Cardano.Ledger.Conway (ApplyTxError (..), ConwayEra)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.PParams
@@ -245,6 +245,9 @@ instance
   , ToExpr (PredicateFailure (EraRule "CERTS" era))
   ) =>
   ToExpr (ConwayLedgerPredFailure era)
+
+instance ToExpr (ApplyTxError ConwayEra) where
+  toExpr (ConwayApplyTxError failures) = toExpr failures
 
 instance ToExpr (Constitution era)