+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/TreeDiff.hs
@@ -23,7 +23,7 @@ import Cardano.Ledger.Alonzo.Plutus.Context (ContextError)
 import Cardano.Ledger.BaseTypes (StrictMaybe)
 import Cardano.Ledger.Binary (EncCBOR (..), FixedSizeCodec (..), natVersion, serialize')
 import qualified Cardano.Ledger.Conway.Rules as Conway
-import Cardano.Ledger.Dijkstra (DijkstraEra)
+import Cardano.Ledger.Dijkstra (ApplyTxError (..), DijkstraEra)
 import Cardano.Ledger.Dijkstra.BlockBody (PerasCert)
 import Cardano.Ledger.Dijkstra.BlockBody.Internal (DijkstraBlockBodyRaw)
 import Cardano.Ledger.Dijkstra.Core (
@@ -241,6 +241,9 @@ instance
 instance
   ToExpr (PredicateFailure (EraRule "LEDGER" era)) =>
   ToExpr (DijkstraMempoolPredFailure era)
+
+instance ToExpr (ApplyTxError DijkstraEra) where
+  toExpr (DijkstraApplyTxError failures) = toExpr failures
 
 instance
   ToExpr (Event (EraRule "LEDGER" era)) =>